🐛 Fix: Open disclosure of active navigation item when link in body is clicked
Previous behaviour: If you were on the Farmlands POS Integration guide and clicked a link in the body to a page under 'References', the 'References' section would show as active but the disclosure would not open.
Change to open the disclosure of the active navigation section when the route changes.

Test plan: Expect disclosure of active navigation section to open when route changes.
